# specs/learning.openapi.yaml
...
openapi: 3.1.0
info:
  title: IT-Shaharcha — Learning Service
  version: 1.0.0
  description: IT skill courses, tutorials, reading docs, and typing practice.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
    description: Versioned platform base path (behind the gateway).
security:
  - bearerAuth: []
...
paths:
...
  /learning/admin/tracks:
    post:
      operationId: createTrack
      summary: Create a skill track.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TrackInput" }
      responses:
        "201": { $ref: "#/components/responses/TrackCreated" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
  /learning/admin/tracks/{trackId}:
    parameters:
      - name: trackId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateTrack
      summary: Update a track.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TrackInput" }
      responses:
        "200": { $ref: "#/components/responses/TrackOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteTrack
      summary: Delete a track.
      tags: [admin]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "409": { $ref: "./shared/common.yaml#/components/responses/Conflict" }
  /learning/admin/courses:
    post:
      operationId: createCourse
      summary: Create a course.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CourseInput" }
      responses:
        "201": { $ref: "#/components/responses/CourseCreated" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
  /learning/admin/courses/{courseId}:
    parameters:
      - name: courseId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateCourse
      summary: Update a course.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CourseInput" }
      responses:
        "200": { $ref: "#/components/responses/CourseOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteCourse
      summary: Delete a course.
      tags: [admin]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
  /learning/admin/modules:
    post:
      operationId: createModule
      summary: Create a module within a course.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModuleInput" }
      responses:
        "201": { $ref: "#/components/responses/ModuleOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
  /learning/admin/modules/{moduleId}:
    parameters:
      - name: moduleId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateModule
      summary: Update a module.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModuleInput" }
      responses:
        "200": { $ref: "#/components/responses/ModuleOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteModule
      summary: Delete a module.
      tags: [admin]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
  /learning/admin/lessons:
    post:
      operationId: createLesson
      summary: Create a lesson within a module.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LessonInput" }
      responses:
        "201": { $ref: "#/components/responses/LessonOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
  /learning/admin/lessons/{lessonId}:
    parameters:
      - name: lessonId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateLesson
      summary: Update a lesson.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LessonInput" }
      responses:
        "200": { $ref: "#/components/responses/LessonOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteLesson
      summary: Delete a lesson.
      tags: [admin]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
  /learning/admin/tutorials:
    post:
      operationId: createTutorial
      summary: Add a tutorial video.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TutorialInput" }
      responses:
        "201": { $ref: "#/components/responses/TutorialOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
  /learning/admin/tutorials/{tutorialId}:
    parameters:
      - name: tutorialId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateTutorial
      summary: Update a tutorial.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TutorialInput" }
      responses:
        "200": { $ref: "#/components/responses/TutorialOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteTutorial
      summary: Delete a tutorial.
      tags: [admin]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
  /learning/admin/docs:
    post:
      operationId: createDoc
      summary: Add a documentation / reading entry.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocInput" }
      responses:
        "201": { $ref: "#/components/responses/DocOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
  /learning/admin/docs/{docId}:
    parameters:
      - name: docId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateDoc
      summary: Update a doc.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocInput" }
      responses:
        "200": { $ref: "#/components/responses/DocOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteDoc
      summary: Delete a doc.
      tags: [admin]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      $ref: "./shared/common.yaml#/components/securitySchemes/bearerAuth"
  schemas:
    Level:
      type: string
      enum: [beginner, intermediate, advanced]
    Track:
      type: object
      required: [id, title]
      properties:
        id: { type: string, format: uuid, example: "33333333-3333-3333-3333-333333333333" }
        title: { type: string, example: "Web Development" }
        slug: { type: string, example: "web-development" }
        description: { type: [string, "null"], example: "From HTML basics to full-stack apps." }
        courseCount: { type: integer, minimum: 0, example: 12 }
    Course:
      type: object
      required: [id, title, level]
      properties:
        id: { type: string, format: uuid, example: "44444444-4444-4444-4444-444444444444" }
        trackId: { type: [string, "null"], format: uuid, example: "33333333-3333-3333-3333-333333333333" }
        title: { type: string, example: "JavaScript Fundamentals" }
        slug: { type: string, example: "javascript-fundamentals" }
        summary: { type: [string, "null"], example: "Variables, functions, and the DOM." }
        level: { $ref: "#/components/schemas/Level" }
        lessonCount: { type: integer, minimum: 0, example: 24 }
        estimatedMinutes: { type: [integer, "null"], minimum: 0, example: 480 }
...
    Module:
      type: object
      required: [id, title, lessons]
      properties:
        id: { type: string, format: uuid, example: "55555555-5555-5555-5555-555555555555" }
        title: { type: string, example: "Functions & Scope" }
        order: { type: integer, minimum: 0, example: 2 }
        lessons:
          type: array
          items: { $ref: "#/components/schemas/Lesson" }
    Lesson:
      type: object
      required: [id, title]
      properties:
        id: { type: string, format: uuid, example: "66666666-6666-6666-6666-666666666666" }
        title: { type: string, example: "Arrow Functions" }
        order: { type: integer, minimum: 0, example: 3 }
        kind:
          type: string
          enum: [reading, video, exercise, quiz]
          default: reading
          example: video
        estimatedMinutes: { type: [integer, "null"], minimum: 0, example: 15 }
...
    Tutorial:
      type: object
      required: [id, title, videoUrl]
      properties:
        id: { type: string, format: uuid, example: "88888888-8888-8888-8888-888888888888" }
        title: { type: string, example: "Building a REST API in Node.js" }
        topic: { type: [string, "null"], example: "backend" }
        videoUrl: { type: string, format: uri, example: "https://videos.itsh.dev/rest-api-node.mp4" }
        durationSeconds: { type: [integer, "null"], minimum: 0, example: 1320 }
        thumbnailUrl: { type: [string, "null"], format: uri, example: "https://cdn.itsh.dev/thumbs/rest-api-node.jpg" }
    Doc:
      type: object
      required: [id, title]
      properties:
        id: { type: string, format: uuid, example: "99999999-9999-9999-9999-999999999999" }
        title: { type: string, example: "HTTP Status Codes Reference" }
        topic: { type: [string, "null"], example: "web" }
        url: { type: [string, "null"], format: uri, example: "https://docs.itsh.dev/http-status-codes" }
        estimatedMinutes: { type: [integer, "null"], minimum: 0, example: 8 }
...
    TrackInput:
      type: object
      required: [title]
      properties:
        title: { type: string, example: "Web Development" }
        slug: { type: [string, "null"], example: "web-development" }
        description: { type: [string, "null"], example: "From HTML basics to full-stack apps." }
    CourseInput:
      type: object
      required: [title, level]
      properties:
        trackId: { type: [string, "null"], format: uuid, example: "33333333-3333-3333-3333-333333333333" }
        title: { type: string, example: "JavaScript Fundamentals" }
        slug: { type: [string, "null"], example: "javascript-fundamentals" }
        summary: { type: [string, "null"], example: "Variables, functions, and the DOM." }
        level: { $ref: "#/components/schemas/Level" }
        estimatedMinutes: { type: [integer, "null"], minimum: 0, example: 480 }
    ModuleInput:
      type: object
      required: [courseId, title]
      properties:
        courseId: { type: string, format: uuid, example: "44444444-4444-4444-4444-444444444444" }
        title: { type: string, example: "Functions & Scope" }
        order: { type: [integer, "null"], minimum: 0, example: 2 }
    LessonInput:
      type: object
      required: [moduleId, title]
      properties:
        moduleId: { type: string, format: uuid, example: "55555555-5555-5555-5555-555555555555" }
        title: { type: string, example: "Arrow Functions" }
        order: { type: [integer, "null"], minimum: 0, example: 3 }
        kind:
          type: string
          enum: [reading, video, exercise, quiz]
          default: reading
          example: video
        estimatedMinutes: { type: [integer, "null"], minimum: 0, example: 15 }
        body: { type: [string, "null"], description: Lesson content (markdown)., example: "# Arrow Functions\n\nArrow functions provide a concise syntax..." }
    TutorialInput:
      type: object
      required: [title, videoUrl]
      properties:
        title: { type: string, example: "Building a REST API in Node.js" }
        topic: { type: [string, "null"], example: "backend" }
        videoUrl: { type: string, format: uri, example: "https://videos.itsh.dev/rest-api-node.mp4" }
        durationSeconds: { type: [integer, "null"], minimum: 0, example: 1320 }
        thumbnailUrl: { type: [string, "null"], format: uri, example: "https://cdn.itsh.dev/thumbs/rest-api-node.jpg" }
        sourceId:
          type: [string, "null"]
          format: uuid
          description: Owning content source, if ingested from a feed.
          example: "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb"
    DocInput:
      type: object
      required: [title]
      properties:
        title: { type: string, example: "HTTP Status Codes Reference" }
        topic: { type: [string, "null"], example: "web" }
        url: { type: [string, "null"], format: uri, example: "https://docs.itsh.dev/http-status-codes" }
        body: { type: [string, "null"], description: Inline doc content (markdown)., example: "## 2xx Success\n\n200 OK — the request succeeded." }
        estimatedMinutes: { type: [integer, "null"], minimum: 0, example: 8 }
        sourceId: { type: [string, "null"], format: uuid, example: "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb" }
...
  responses:
    TrackCreated: &trackBody
      description: The track.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data: { $ref: "#/components/schemas/Track" }
...
    CourseCreated: &courseBody
      description: The course.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data: { $ref: "#/components/schemas/Course" }
...
    ModuleOk:
      description: The module.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data: { $ref: "#/components/schemas/Module" }
    LessonOk:
      description: The lesson.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data: { $ref: "#/components/schemas/Lesson" }
    TutorialOk:
      description: The tutorial.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data: { $ref: "#/components/schemas/Tutorial" }
    DocOk:
      description: The doc.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data: { $ref: "#/components/schemas/Doc" }
...